+++ docs/html/_sensor_manager_8cpp.js
@@ -1,5 +1,8 @@
 var _sensor_manager_8cpp =
 [
+    [ "anonymous_namespace{SensorManager.cpp}::batterySampleLooksSuspicious", "namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#ab5a3fd276944117f8d63d688914a47d4", null ],
+    [ "anonymous_namespace{SensorManager.cpp}::estimateSocFromVoltage", "namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#a5e3e975006a3ea6d5ce637eef262bc84", null ],
     [ "anonymous_namespace{SensorManager.cpp}::probeTmp112Present", "namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#a452bb1633016be3c71fffc2fe5d5a0d7", null ],
-    [ "batteryContext", "_sensor_manager_8cpp.html#aef74f36dc982b031c818fa1eaed04c92", null ]
+    [ "batteryContext", "_sensor_manager_8cpp.html#aef74f36dc982b031c818fa1eaed04c92", null ],
+    [ "fuelGauge", "_sensor_manager_8cpp.html#a8858da7e857adf69b17865a3ab3d322f", null ]
 ];