+++ docs/html/search/all_5.js
@@ -2,7 +2,8 @@ var searchData=
 [
   ['ensuresensorenabled_0',['ensureSensorEnabled',['../_state___common_8h.html#a26af5a78d8cca160ca913b5e91a224d6',1,'ensureSensorEnabled(const char *context):&#160;State_Idle.cpp'],['../_state___idle_8cpp.html#a26af5a78d8cca160ca913b5e91a224d6',1,'ensureSensorEnabled(const char *context):&#160;State_Idle.cpp']]],
   ['error_5fstate_1',['ERROR_STATE',['../_state_machine_8h.html#a5d74787dedbc4e11c1ab15bf487e61f8aa07813f3418b83dc769daa78689fb9c3',1,'StateMachine.h']]],
-  ['example_2',['Multiple Products Example',['../index.html#autotoc_md28',1,'']]],
-  ['examples_3',['Webhook Resolution Examples',['../index.html#autotoc_md26',1,'']]],
-  ['externaltempc_4',['externalTempC',['../classcurrent_status_data_1_1_current_data.html#a6631f1c5e492a839917a32ab8a5d9ce5',1,'currentStatusData::CurrentData']]]
+  ['estimatesocfromvoltage_2',['estimateSocFromVoltage',['../namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#a5e3e975006a3ea6d5ce637eef262bc84',1,'anonymous_namespace{SensorManager.cpp}']]],
+  ['example_3',['Multiple Products Example',['../index.html#autotoc_md28',1,'']]],
+  ['examples_4',['Webhook Resolution Examples',['../index.html#autotoc_md26',1,'']]],
+  ['externaltempc_5',['externalTempC',['../classcurrent_status_data_1_1_current_data.html#a6631f1c5e492a839917a32ab8a5d9ce5',1,'currentStatusData::CurrentData']]]
 ];