+++ docs/html/search/all_c.js
@@ -1,7 +1,10 @@
 var searchData=
 [
   ['name_0',['name',['../struct_sensor_definition.html#ad5557bac593d6f23d8e0931a24ed0172',1,'SensorDefinition']]],
-  ['none_1',['NONE',['../struct_observability_1_1_wake_cycle_stats.html#a54f787bfc1fab38f76cbd367716e8c12ab50339a10e1de285ac99d4c3990b8693',1,'Observability::WakeCycleStats']]],
-  ['normal_2',['NORMAL',['../struct_observability_1_1_wake_cycle_stats.html#a54f787bfc1fab38f76cbd367716e8c12a1e23852820b9154316c7c06e2b7ba051',1,'Observability::WakeCycleStats']]],
-  ['not_5fattempted_3',['NOT_ATTEMPTED',['../struct_observability_1_1_wake_cycle_stats.html#aba2be75384b01e768f8572821ec378b3a44351b1ee085a81f2c6fe6fb26d6e379',1,'Observability::WakeCycleStats']]]
+  ['nightly_5fheap_5freset_5fthreshold_5fbytes_1',['NIGHTLY_HEAP_RESET_THRESHOLD_BYTES',['../namespace_connectivity_policy.html#ae00914c08bf3eaaa5d8e412d63d07078',1,'ConnectivityPolicy']]],
+  ['nightly_5fheap_5fwarn_5fthreshold_5fbytes_2',['NIGHTLY_HEAP_WARN_THRESHOLD_BYTES',['../namespace_connectivity_policy.html#a264374785dc6d61fd971454e4c9a0e7e',1,'ConnectivityPolicy']]],
+  ['none_3',['NONE',['../struct_observability_1_1_wake_cycle_stats.html#a54f787bfc1fab38f76cbd367716e8c12ab50339a10e1de285ac99d4c3990b8693',1,'Observability::WakeCycleStats']]],
+  ['normal_4',['NORMAL',['../struct_observability_1_1_wake_cycle_stats.html#a54f787bfc1fab38f76cbd367716e8c12a1e23852820b9154316c7c06e2b7ba051',1,'Observability::WakeCycleStats']]],
+  ['not_5fattempted_5',['NOT_ATTEMPTED',['../struct_observability_1_1_wake_cycle_stats.html#aba2be75384b01e768f8572821ec378b3a44351b1ee085a81f2c6fe6fb26d6e379',1,'Observability::WakeCycleStats']]],
+  ['notewakefromlowpowersleep_6',['noteWakeFromLowPowerSleep',['../class_sensor_manager.html#ac51dedb8c2fcb06c5f5839b3e9e84ea9',1,'SensorManager']]]
 ];