+++ docs/html/search/functions_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['batterystate_0',['batteryState',['../class_sensor_manager.html#adabba3bce18318e591c43645fadf833c',1,'SensorManager']]]
+  ['batterysamplelookssuspicious_0',['batterySampleLooksSuspicious',['../namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#ab5a3fd276944117f8d63d688914a47d4',1,'anonymous_namespace{SensorManager.cpp}']]],
+  ['batterystate_1',['batteryState',['../class_sensor_manager.html#adabba3bce18318e591c43645fadf833c',1,'SensorManager']]]
 ];