+++ docs/html/search/functions_12.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['writedevicestatustocloud_0',['writeDeviceStatusToCloud',['../class_cloud.html#a27758d9343f1531651c739f13ff6006b',1,'Cloud']]]
+  ['validate_0',['validate',['../classsys_status_data.html#af98d0dc49eaafea2cf962d0d9d49e505',1,'sysStatusData::validate()'],['../classsensor_config_data.html#a7cad90568ca2b028283b5f80542d2a4f',1,'sensorConfigData::validate()'],['../classcurrent_status_data.html#ab677c0444b73ce95c05d4692763fb7a8',1,'currentStatusData::validate()']]]
 ];