+++ docs/html/search/functions_d.js
@@ -1,12 +1,11 @@
 var searchData=
 [
-  ['raisealert_0',['raiseAlert',['../classcurrent_status_data.html#a3b9d02e3ef08b2a9629ff053aacee27c',1,'currentStatusData']]],
-  ['readtmp112temperaturec_1',['readTmp112TemperatureC',['../class_sensor_manager.html#a87645f59d0a2fe8c6644b386f24b7692',1,'SensorManager']]],
-  ['recordstatetransition_2',['recordStateTransition',['../class_thrash_guard.html#a363e2fac16a9fad4bdebc0f54bea80a0',1,'ThrashGuard']]],
-  ['requestclouddisconnectonly_3',['requestCloudDisconnectOnly',['../namespace_connectivity.html#a0e2777379673ce5006c16847bc640547',1,'Connectivity']]],
-  ['requestfulldisconnectandradiooff_4',['requestFullDisconnectAndRadioOff',['../namespace_connectivity.html#a50bb34193371f247ff8639d8fba69409',1,'Connectivity']]],
-  ['requestradiopoweroff_5',['requestRadioPowerOff',['../namespace_connectivity.html#a2a57cc42c82eb80256258e783f248508',1,'Connectivity']]],
-  ['reset_6',['reset',['../class_i_sensor.html#a5a11e705590a9a16bef6c0dd154da2e7',1,'ISensor::reset()'],['../class_p_i_r_sensor.html#a0f0d1a8a3d010e2fbef7acaed0b381b7',1,'PIRSensor::reset()']]],
-  ['reseteverything_7',['resetEverything',['../classcurrent_status_data.html#a338c2e9242f5cbbc12a1be90861c45cb',1,'currentStatusData']]],
-  ['resetonwake_8',['resetOnWake',['../struct_observability_1_1_wake_cycle_stats.html#a06cab85b425af73192f92487cb60d163',1,'Observability::WakeCycleStats']]]
+  ['particle_5ffunctions_0',['Particle_Functions',['../class_particle___functions.html#af4f9704a1c5bdd9983d7803ef53d8894',1,'Particle_Functions::Particle_Functions()'],['../class_particle___functions.html#a5e1c905daf41ddd01790ccd384310285',1,'Particle_Functions::Particle_Functions(const Particle_Functions &amp;)=delete']]],
+  ['probetmp112present_1',['probeTmp112Present',['../namespaceanonymous__namespace_02_sensor_manager_8cpp_03.html#a452bb1633016be3c71fffc2fe5d5a0d7',1,'anonymous_namespace{SensorManager.cpp}']]],
+  ['product_5fversion_2',['PRODUCT_VERSION',['../_generalized-_core-_counter_8cpp.html#ad500242a86166c1fe064fd0500bda870',1,'Generalized-Core-Counter.cpp']]],
+  ['publishdata_3',['publishData',['../_generalized-_core-_counter_8cpp.html#a6e8d6c80794e3a266d4ef6b976e9c27c',1,'publishData():&#160;Generalized-Core-Counter.cpp'],['../_state___common_8h.html#a6e8d6c80794e3a266d4ef6b976e9c27c',1,'publishData():&#160;Generalized-Core-Counter.cpp'],['../_state_machine_8h.html#a6e8d6c80794e3a266d4ef6b976e9c27c',1,'publishData():&#160;Generalized-Core-Counter.cpp']]],
+  ['publishdatatoledger_4',['publishDataToLedger',['../class_cloud.html#a832b2fee272f68202c5877a0b2fee7c0',1,'Cloud']]],
+  ['publishdiagnosticsafe_5',['publishDiagnosticSafe',['../_generalized-_core-_counter_8cpp.html#a5b6d4035a9304bc53863ee513f708b0d',1,'publishDiagnosticSafe(const char *eventName, const char *data, PublishFlags flags):&#160;Generalized-Core-Counter.cpp'],['../_my_persistent_data_8cpp.html#ab7e9803d96548944e5e40d23163484c6',1,'publishDiagnosticSafe(const char *eventName, const char *data, PublishFlags flags=PRIVATE):&#160;Generalized-Core-Counter.cpp'],['../_state___common_8h.html#a5b6d4035a9304bc53863ee513f708b0d',1,'publishDiagnosticSafe(const char *eventName, const char *data, PublishFlags flags):&#160;Generalized-Core-Counter.cpp']]],
+  ['publishstartupstatus_6',['publishStartupStatus',['../_generalized-_core-_counter_8cpp.html#a3b52d4e2736530265377e967979feaf9',1,'Generalized-Core-Counter.cpp']]],
+  ['publishstatetransition_7',['publishStateTransition',['../_generalized-_core-_counter_8cpp.html#afb2932b86bac021e349213b370798543',1,'publishStateTransition():&#160;Generalized-Core-Counter.cpp'],['../_state_machine_8h.html#afb2932b86bac021e349213b370798543',1,'publishStateTransition():&#160;Generalized-Core-Counter.cpp']]]
 ];