+++ docs/html/search/searchdata.js
@@ -4,7 +4,7 @@ var indexSectionsWithContent =
   1: "cilpstw",
   2: "aclops",
   3: "bcdgilmprstvw",
-  4: "abcdefghilmoprstuvw~",
+  4: "abcdefghilmnoprstuvw~",
   5: "_abcdefghilmnoprstuvw",
   6: "abcrs",
   7: "abcdefilmnoprstv",