+++ docs/html/search/variables_c.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['name_0',['name',['../struct_sensor_definition.html#ad5557bac593d6f23d8e0931a24ed0172',1,'SensorDefinition']]]
+  ['name_0',['name',['../struct_sensor_definition.html#ad5557bac593d6f23d8e0931a24ed0172',1,'SensorDefinition']]],
+  ['nightly_5fheap_5freset_5fthreshold_5fbytes_1',['NIGHTLY_HEAP_RESET_THRESHOLD_BYTES',['../namespace_connectivity_policy.html#ae00914c08bf3eaaa5d8e412d63d07078',1,'ConnectivityPolicy']]],
+  ['nightly_5fheap_5fwarn_5fthreshold_5fbytes_2',['NIGHTLY_HEAP_WARN_THRESHOLD_BYTES',['../namespace_connectivity_policy.html#a264374785dc6d61fd971454e4c9a0e7e',1,'ConnectivityPolicy']]]
 ];